"CPU it's over 9000!" resolved
The problem was (probably!) resolved by adding a return statement in the while cycle in ClientRequestThread class. If the client request is a null string, then the run method returns to the main. But the thread is still alive: to check this you can run the server and then the Test03 in tests/performance_tests. It registered N users, and then login these users: every user has a personal socket used both for the registration and the login. Thanks to the debug println in StartServer, that prints the ID of the thread every time a thread was started, you can see that there are exactly N threads (and not N*2 thread). And the CPU is happy.